feat: 升级项目依赖到最新稳定版本并优化动态权重精度
依赖升级：
- 升级 Go 依赖：sonic v1.15.0, mimetype v1.4.13, redis v9.17.3
- 升级 Vue 依赖：axios v1.13.4, vue-router v5.0.2, @types/node v25.2.0, prettier v3.8.1
- 升级前端开发依赖：@vitejs/plugin-vue v6.0.4, @vueuse/core v14.2.0, vue-tsc v3.2.4

功能优化：
- 优化动态权重计算精度，从整数改为保留1位小数的浮点数
- 更新有效权重最小值从1调整为0.1，提供更精细的权重控制
- 改进权重显示格式，统一使用1位小数展示

代码改进：
- 优化子组权重编辑模态框的数值处理逻辑
- 改进组表单模态框的权重验证和显示
- 增强子组表格的权重格式化显示功能
- 更新类型定义以支持浮点数权重值
- 完善动态权重管理器的测试用例覆盖

脚本优化：
- 改进 PGO 性能分析脚本的错误处理和输出格式

国际化：
- 补充日语翻译文本

测试增强：
- 扩展动态权重测试用例，覆盖浮点数精度场景
- 增加边界值和精度验证测试

+++ internal/models/types.go
@@ -118,7 +118,7 @@ type SubGroupInfo struct {
 type DynamicWeightInfo struct {
 	BaseWeight      int     `json:"base_weight"`      // Original configured weight
 	HealthScore     float64 `json:"health_score"`     // Health score (0.0 - 1.0)
-	EffectiveWeight int     `json:"effective_weight"` // Calculated effective weight
+	EffectiveWeight float64 `json:"effective_weight"` // Calculated effective weight (1 decimal place, min 0.1)
 	SuccessRate     float64 `json:"success_rate"`     // Success rate percentage (0-100)
 	RequestCount    int64   `json:"request_count"`    // Total request count
 	LastFailureAt   *string `json:"last_failure_at"`  // Last failure timestamp (ISO8601, nil if never failed)
